Display Error When Playing A Round
When a player is able to make a move, player can select a square on computer's board. If a square hit, it turns red. If not, it turns green. After being clicked, said square will delete the event listener for when mouse is hovered and mouse has left that square, letting it remain either red or green since it is already clicked. However, event listener seems not to be removed and when mouse left said square, it still remove its colour